new branch to focus on refactoring API and adding new stubs for OpenAI endpoints
