+++ src/leapfrogai_api/Makefile
@@ -1,0 +1,6 @@
+install:
+	python -m pip install -e .
+
+dev:
+	make install
+	python -m uvicorn main:app --port 3000 --reload
