+++ src/leapfrogai_api/pyproject.toml
@@ -11,7 +11,6 @@ dependencies = [
     "python-multipart >= 0.0.7", #indirect dep of FastAPI to receive form data for file uploads
     "watchfiles >= 0.21.0",
     "leapfrogai_sdk",
-    "supabase",
 ]
 requires-python = "~=3.11"
